back-end/requirements.txt
absl-py==0.1.13
astor==0.6.2
bleach==1.5.0
boto3==1.6.16
botocore==1.9.16
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
docutils==0.14
Flask==0.12.2
Flask-Cors==3.0.3
Flask-SocketIO==2.9.6
gast==0.2.0
google-api-core==1.0.0
google-auth==1.4.1
google-cloud-language==1.0.1
google-cloud-speech==0.32.1
googleapis-common-protos==1.5.3
graphviz==0.8.2
grpcio==1.10.0
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
Markdown==2.6.11
MarkupSafe==1.0
nltk==3.2.5
numpy==1.14.2
Pillow==5.1.0
protobuf==3.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pygraphviz==1.3.1
python-dateutil==2.6.1
python-engineio==2.0.4
python-socketio==1.9.0
pytz==2018.3
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
tflearn==0.3.2
transitions==0.6.4
urllib3==1.22
Werkzeug==0.14.1
